+++ longRange.hetero/profile.test/template/tools/scripts/spme.ik.esti.sh
@@ -2,6 +2,11 @@
 
 source parameters.sh
 
+if test ! -d $record_dir; then
+    echo "no record dir $record_dir"
+    exit
+fi
+
 mylog=spme.ik.esti.log
 rm -f $mylog
 touch $mylog
@@ -23,5 +28,4 @@ mv -f meanf.out $record_dir/esti.rec.spme.ik.meanf.out
 # combine rec with dir
 ./tools/analyze/combine.error --dir-meanf $record_dir/esti.dir.meanf.out --rec-meanf $record_dir/esti.rec.spme.ik.meanf.out  --dir-error $record_dir/esti.dir.error.out --rec-error $record_dir/esti.rec.spme.ik.error.out --output-error $record_dir/esti.spme.ik.error.out --output-meanf $record_dir/esti.spme.ik.meanf.out &>> $mylog
 
-
-mv -f $mylog $record_dir
+mv -f make.log $mylog $record_dir
